pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "wanikani_apprentice"
authors = [{name = "Andy Freeland", email = "andy@andyfreeland.net"}]
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
requires-python = ">=3.10"

dependencies = [
  "attrs",
  "ciso8601",
  "httpx",
]

[project.urls]
Home = "https://github.com/rouge8/wanikani-apprentice"

[tool.mypy]
strict = true
allow_redefinition = true
